Reuse Launcher orchestration in EngineTestKit
